crypto: moving core code to core
