+++ source/common/crypto/crypto_impl.cc
@@ -1,4 +1,4 @@
-#include "source/extensions/common/crypto/crypto_impl.h"
+#include "source/common/crypto/crypto_impl.h"
 
 namespace Envoy {
 namespace Common {
