+++ source/extensions/all_extensions.bzl
@@ -4,7 +4,6 @@ load("@envoy_build_config//:extensions_build_config.bzl", "EXTENSIONS")
 # These extensions are registered using the extension system but are required for the core Envoy build.
 # The map may be overridden by extensions specified in envoy_build_config.
 _required_extensions = {
-    "envoy.common.crypto.utility_lib": "//source/extensions/common/crypto:utility_lib",
     "envoy.http.original_ip_detection.xff": "//source/extensions/http/original_ip_detection/xff:config",
     "envoy.request_id.uuid": "//source/extensions/request_id/uuid:config",
     "envoy.transport_sockets.tls": "//source/extensions/transport_sockets/tls:config",
