+++ source/extensions/filters/http/lua/lua_filter.cc
@@ -9,6 +9,7 @@
 #include "source/common/common/assert.h"
 #include "source/common/common/enum_to_int.h"
 #include "source/common/config/datasource.h"
+#include "source/common/crypto/crypto_impl.h"
 #include "source/common/crypto/utility.h"
 #include "source/common/http/message_impl.h"
