+++ test/per_file_coverage.sh
@@ -22,7 +22,6 @@ declare -a KNOWN_LOW_COVERAGE=(
 "source/common/watchdog:42.9" # Death tests don't report LCOV
 "source/common/config/xds_mux:94.5"
 "source/exe:94.3"
-"source/extensions/common/crypto:91.5"
 "source/extensions/common/tap:95.9"
 "source/extensions/common/wasm:95.3" # flaky: be careful adjusting
 "source/extensions/common/wasm/null:77.8"
